pyproject.toml
[tool.poetry]
name = "streamcontroller"
version = "1.0.0"
description = ""
authors = ["b0wter <b0wter@gmx.de>"]

[tool.poetry.dependencies]
python = "^3.7"
streamdeck = "^0.7.3"
wheel = "^0.34.2"
pillow = "^7.1.1"
pynput = "^1.6.8"
xlib = "^0.21"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.7.9"
black = {version = "^19.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
